Add BigDecimalUtils arch test

+++ src/main/resources/db/migration/V4__add_stock_item_table.sql
@@ -6,6 +6,6 @@ START WITH 1
 CREATE TABLE IF NOT EXISTS stock_items (
     id INTEGER NOT NULL DEFAULT nextval('stock_item_sequence') PRIMARY KEY,
     purchase_order_line_id INTEGER,
-    quantity INTEGER,
+    quantity DECIMAL(10, 2),
     CONSTRAINT fk_purchase_order_line FOREIGN KEY (purchase_order_line_id) REFERENCES purchase_order_lines(id)
 );
